requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
pydantic
passlib[argon2]
python-jose
httpx
pytest